pyproject.toml
[project]
name = "NavTradev1"
version = "1.0.0"
description = "Document processing pipeline with LLM chat interface"
requires-python = ">=3.10"
dependencies = [
    # FastAPI backend
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.0.0",
    # Streamlit frontend
    "streamlit>=1.28.0",
    "requests>=2.31.0",
    # Document conversion
    "docling>=2.0.0",
    "docling-hierarchical-pdf>=0.1.0",
    "langchain-text-splitters>=0.3.0",
    "playwright>=1.40.0",
    # Utilities
    "python-multipart>=0.0.6",
    "python-dotenv>=1.2.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[tool.black]
line-length = 100
target-version = ['py310']

[tool.ruff]
line-length = 100
target-version = "py310"
